refactor: move all plugins to the lua/plugins dir, refactor lazy bootstrap, move keymaps & autocmds to config dir
