+++ .config/fish/functions/wezterm.fish
@@ -1,0 +1,4 @@
+function wezterm --wraps='flatpak run org.wezfurlong.wezterm' --description 'alias wezterm=flatpak run org.wezfurlong.wezterm'
+  flatpak run org.wezfurlong.wezterm $argv
+        
+end
