+++ .config/flameshot/flameshot.ini
@@ -1,5 +1,6 @@
 [General]
 contrastOpacity=188
+drawColor=#ffff00
 savePath=/home/nathan/Pictures/screenshots
 showStartupLaunchMessage=false
 startupLaunch=true
