--- .config/nvim/after/plugin/autotag.lua
@@ -1,7 +1,0 @@
-require('nvim-ts-autotag').setup {
-  opts = {
-    enable_close = true, -- Auto close tags
-    enable_rename = true, -- Auto rename pairs of tags
-    enable_close_on_slash = true, -- Auto close on trailing </
-  },
-}
