--- .config/nvim/after/plugin/textcase.lua
@@ -1,1 +1,0 @@
-require('textcase').setup {}
