--- .config/nvim/after/plugin/ufo.lua
@@ -1,1 +1,0 @@
-require("ufo").setup()
