+++ .config/nvim/init.lua
@@ -1,8 +1,7 @@
-require("options")
-require("bootstrap")
-require("plugins")
-require("lsp")
-require("keymaps")
+require("config.options")
+require("config.keymaps")
+require("config.lazy")
+require("config.autocmd")
 
 vim.api.nvim_set_hl(0, "ColorColumn", { bg = "#363433" })
