+++ .config/nvim/lua/config/autocmd.lua
@@ -1,0 +1,5 @@
+vim.api.nvim_create_autocmd({ "BufReadPost" }, {
+	desc = "Open file at the last position it was opened",
+	pattern = "*",
+	command = 'silent! normal! g`"zv',
+})
