+++ .config/nvim/lua/config/lazy/init.lua
@@ -1,0 +1,32 @@
+local lazy_utils = require("config.lazy.utils")
+lazy_utils.bootstrap()
+
+-- load plugins
+require("lazy").setup({
+	spec = lazy_utils.generate_import_specs(),
+}, {
+	checker = {
+		enabled = true,
+		notify = false,
+	},
+	change_detection = {
+		notify = false,
+	},
+	ui = {
+		icons = vim.g.have_nerd_font and {} or {
+			cmd = '⌘',
+			config = '🛠',
+			event = '📅',
+			ft = '📂',
+			init = '⚙',
+			keys = '🗝',
+			plugin = '🔌',
+			runtime = '💻',
+			require = '🌙',
+			source = '📄',
+			start = '🚀',
+			task = '📌',
+			lazy = '💤 ',
+		},
+	},
+})
