+++ .config/nvim/lua/plugins/autopairs.lua
@@ -1,0 +1,1 @@
+return { 'windwp/nvim-autopairs', event = 'InsertEnter', config = true, opts = {} }
