+++ .config/nvim/lua/plugins/bqf.lua
@@ -1,0 +1,1 @@
+return { 'kevinhwang91/nvim-bqf' }
