+++ .config/nvim/lua/plugins/cmp.lua
@@ -1,0 +1,21 @@
+return {
+  'hrsh7th/nvim-cmp',
+  dependencies = {
+    'L3MON4D3/LuaSnip',
+    'saadparwaiz1/cmp_luasnip',
+    'hrsh7th/cmp-nvim-lsp',
+    'hrsh7th/cmp-path',
+    'luckasRanarison/tailwind-tools.nvim',
+    'onsails/lspkind-nvim',
+  },
+  opts = function()
+    return {
+      -- ...
+      formatting = {
+        format = require("lspkind").cmp_format({
+          before = require("tailwind-tools.cmp").lspkind_format
+        }),
+      },
+    }
+  end,
+}
