+++ .config/nvim/lua/plugins/comment.lua
@@ -1,0 +1,6 @@
+return {
+    'terrortylor/nvim-comment',
+    config = function()
+        require("nvim_comment").setup({})
+    end
+}
