+++ .config/nvim/lua/plugins/git-blame.lua
@@ -1,0 +1,1 @@
+return { 'f-person/git-blame.nvim' }
