+++ .config/nvim/lua/plugins/gruvbox.lua
@@ -1,0 +1,12 @@
+return {
+  'sainnhe/gruvbox-material',
+  lazy = false,
+  priority = 999,
+  config = function()
+    -- Optionally configure and load the colorscheme
+    -- directly inside the plugin declaration.
+    vim.g.gruvbox_material_enable_italic = true
+    vim.g.gruvbox_material_background = 'soft'
+    vim.cmd.colorscheme 'gruvbox-material'
+  end,
+}
