+++ .config/nvim/lua/plugins/lazygit.lua
@@ -1,0 +1,19 @@
+return {
+  'kdheepak/lazygit.nvim',
+  cmd = {
+    'LazyGit',
+    'LazyGitConfig',
+    'LazyGitCurrentFile',
+    'LazyGitFilter',
+    'LazyGitFilterCurrentFile',
+  },
+  -- optional for floating window border decoration
+  dependencies = {
+    'nvim-lua/plenary.nvim',
+  },
+  -- setting the keybinding for LazyGit with 'keys' is recommended in
+  -- order to load the plugin when the command is run for the first time
+  keys = {
+    { '<leader>lg', '<cmd>LazyGit<cr>', desc = 'LazyGit' }
+  }
+}
