+++ .config/nvim/lua/plugins/markview.lua
@@ -1,0 +1,4 @@
+return {
+  "OXY2DEV/markview.nvim",
+  lazy = false,
+};
