+++ .config/nvim/lua/plugins/netrw.lua
@@ -1,7 +1,3 @@
-require('netrw').setup {
-  use_devicons = true,
-}
-
 vim.g.netrw_altfile = 1
 
 -- vim.keymap.set("n", "<leader>", { desc = "" })
@@ -13,3 +9,8 @@ vim.keymap.set(
   ':Ex <bar> :sil! /<C-R>=expand("%:t")<CR><CR>',
   { desc = '[P]roject [E]xplore' }
 )
+
+return { 'prichrd/netrw.nvim', opts = {
+  use_devicons = true,
+ }
+ }
