+++ .config/nvim/lua/plugins/spectre.lua
@@ -1,0 +1,6 @@
+return {
+  'nvim-pack/nvim-spectre',
+  dependencies = {
+    'nvim-lua/plenary.nvim'
+  }
+}
