+++ .config/nvim/lua/plugins/tailwind-tools.lua
@@ -1,0 +1,9 @@
+return {
+  "luckasRanarison/tailwind-tools.nvim",
+  name = "tailwind-tools",
+  build = ":UpdateRemotePlugins",
+  dependencies = {
+    "nvim-treesitter/nvim-treesitter",
+  },
+  opts = {}
+}
