+++ .config/nvim/lua/plugins/text-case.lua
@@ -1,0 +1,5 @@
+-- Replace text by case, i.e camelCase to snake_case
+return {
+  'johmsalas/text-case.nvim',
+opts = {}
+}
