+++ .config/nvim/lua/plugins/vim-sleuth.lua
@@ -1,0 +1,1 @@
+return { 'tpope/vim-sleuth' }
