+++ .config/nvim/lua/plugins/vim-surround.lua
@@ -1,0 +1,1 @@
+return { 'tpope/vim-surround' }
