+++ .config/nvim/lua/plugins/which-key.lua
@@ -1,0 +1,1 @@
+return { 'folke/which-key.nvim', opts = {} }
